requirements.txt
pandas==1.1.5
numpy==1.18.5
numba==0.53.1
scikit-learn==0.23.1
tensorflow-gpu==1.13.1
tqdm==4.46.1
scipy==1.5.0
deepctr==0.8.6
gensim==3.8